refactor: add initial main class for DECIDE()
App.java and AppTest.java were two files generated by the maven
generate command. In order to have something more suitable for the
program, these two files is now replaced with the main class that has
the DECIDE() function, LaunchInterceptor.java

DECIDE() itself is missing implementation and is empty. This has to be
implemented later.